fix(beefy): convert token amount to shares amount for withdrawal (#818)
### Description

Beefy Valut V7 expects the amount of shares, and not the amount of
underlying token, to be provided for the `withdraw()` method. This
requires us to convert the token amount to a shares amount inside the
shortcut to ensure the correct amount is withdrawn.

*
[docs](https://docs.beefy.finance/developer-documentation/vault-contract#withdraw)
* [contract
code](https://github.com/beefyfinance/beefy-contracts/blob/5b3f00395d70e65f3f246dce30f1219254e8d04d/contracts/BIFI/vaults/BeefyVaultV7.sol#L139-L155)

### TODO

There is clearly a race condition between reading the contract data and
executing the withdrawal, which may affect the precision of the
operation. We might want to use a helper contract to encapsulate all
readings within a single transaction.